chore(SimuResult): separate files
